+++ my-app/src/components/pages/numSequences/NumSequences.js
