+++ my-app/src/components/pages/passwordGen/PasswordGen.js
